Character Shadow - Bounding Box Glitches https://jira.ccpgames.com/browse/EVEDEV-8330
- fixed compile error
- added using dynamic bounding box for skinned objects
- cleaned up debug rendering in skinned objects
- fixed "display" check box for skinned objects

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/Trinity/": change = 1564561]

+++ Tr2MeshBase.cpp
@@ -204,7 +204,7 @@ bool Tr2MeshBase::GetDynamicBoundingBox( const Matrix* boneTransforms, Vector3& 
 
 	BoundingBoxInitialize( min, max );
 
-	for( int i = 0; i < m_jointMappingAnimRig.size(); ++i )
+	for( size_t i = 0; i < m_jointMappingAnimRig.size(); ++i )
 	{
 		auto& joint = meshData->m_jointBindings[i];
 		const Matrix& m = boneTransforms[m_jointMappingAnimRig[i]];
